cbmem: replace pointer type by uint64_t
Since coreboot is compiled into 32bit code, and userspace
might be 32 or 64bit, putting a pointer into the coreboot
table is not viable. Instead, use a uint64_t, which is always
big enough for a pointer, even if we decide to move to a 64bit
coreboot at some point.

Reviewed-on: http://review.coreboot.org/2135
Tested-by: build bot (Jenkins)

+++ src/arch/armv7/boot/coreboot_table.c
@@ -251,7 +251,7 @@ static void add_cbmem_pointers(struct lb_header *header)
 		}
 		cbmem_ref->tag = sid->table_tag;
 		cbmem_ref->size = sizeof(*cbmem_ref);
-		cbmem_ref->cbmem_addr = cbmem_addr;
+		cbmem_ref->cbmem_addr = (unsigned long)cbmem_addr;
 	}
 }
